proyectos_alternativos_1/proyectos_alternativos/chatbot_directorio/requirements.txt
sentence-transformers
scikit-learn
numpy
fastapi
uvicorn[standard]
python-dotenv
